Refactor upgrade structure
