+++ upgrade/README.md
@@ -21,7 +21,7 @@ cp .env.example .env
 Complete the variables in `.env` file. The `ROLE` variable indicates if the validator will send the creation transaction on the multisig wallet or if it will send the confirmation.
 
 Run the script. The following are available:
-* `npm run upgradeBridgeOnHome`
-* `npm run upgradeValidatorsOnForeign`
-* `npm run upgradeBridgeOnForeign`
-* `npm run migrateToMCD`
+* `npm run mcdUpgrade:upgradeBridgeOnHome`
+* `npm run mcdUpgrade:upgradeValidatorsOnForeign`
+* `npm run mcdUpgrade:upgradeBridgeOnForeign`
+* `npm run mcdUpgrade:migrateToMCD`
