+++ upgrade/src/mcdUpgrade/migrateToMCD.js
@@ -1,10 +1,10 @@
 require('dotenv').config()
 const Web3 = require('web3')
-const multiSigWalletAbi = require('../abi/multiSigwallet')
-const proxyAbi = require('../../build/contracts/EternalStorageProxy').abi
-const foreignBridgeAbi = require('../../build/contracts/ForeignBridgeErcToNative').abi
-const confirmTransaction = require('./confirmTransaction')
-const validatorState = require('./validatorState')
+const multiSigWalletAbi = require('../../abi/multiSigwallet')
+const proxyAbi = require('../../../build/contracts/EternalStorageProxy').abi
+const foreignBridgeAbi = require('../../../build/contracts/ForeignBridgeErcToNative').abi
+const confirmTransaction = require('../utils/confirmTransaction')
+const validatorState = require('../utils/validatorState')
 
 const {
   FOREIGN_PRIVKEY,
